// ui/src/lib/booking-hotel-api/model/city-model.ts
/**
 * main-app API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CityModel { 
    id: number;
    name: string;
    countryId?: number;
    createdAt: string;
    updatedAt: string;
    removed: boolean;
}