// ui/src/lib/booking-hotel-api/model/country-model.ts
/**
 * main-app API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CountryModel { 
    id: number;
    name: string;
    currency: string;
    code: string;
    createdAt: string;
    updatedAt: string;
    removed: boolean;
}